refactor: lift and shift trnsactions and master
